User Scripts: update debug-kmsg.sh to use BYO busybox

+++ sdcard_tweaks/roms/MAME/debug-kmsg.sh
@@ -15,22 +15,31 @@ cat /dev/urandom > /dev/fb0
 
 # Lets try catting /dev/kmsg in the background, sleeping for a while and then killing it gently.
 
-log "Running cat in the background"
-cat /dev/kmsg > "${output_dir}/debug-kmsg.txt" 2>&1 & 
-cat_pid=$!
+if [ ! -f "${byobb}" ]; then
+    log "The built-in busybox doesn't have this command."
+    log "Try: https://github.com/EXALAB/Busybox-static/tree/main/busybox_arm"
+    log "Put it in: ${GT4286UTIL_HOME}/bin/busybox_arm"
+else
+    log "BYO Busybox found: ${byobb}"
 
-# kill %1 #The first job
-# kill %- #The last job
-# kill $! #As long as you don't launch any other process in the background you can use $! directly
 
-log "Sleeping for 5"
-sleep 5
-/bin/sync
-log "Kill cat with TERM"
-kill -TERM ${cat_pid}
-/bin/sync
-sleep 5
-log "Kill cat with 9"
-kill -9 ${cat_pid}
-/bin/sync
-log "Done"
+    log "Running cat in the background"
+    cat /dev/kmsg > "${output_dir}/debug-kmsg.txt" 2>&1 & 
+    cat_pid=$!
+
+    # kill %1 #The first job
+    # kill %- #The last job
+    # kill $! #As long as you don't launch any other process in the background you can use $! directly
+
+    log "Sleeping for 5"
+    ${byobb} sleep 5
+    /bin/sync
+    log "Kill cat with TERM"
+    kill -TERM ${cat_pid}
+    /bin/sync
+    ${byobb} sleep 5
+    log "Kill cat with 9"
+    kill -9 ${cat_pid}
+    /bin/sync
+    log "Done"
+fi
